[refactor] Rename utility imports from splat_tb_utils to splatTableUtils and gallery_tb_utils to galleryTableUtils; update related API routes
