+++ src/app/api/galleries/gallery/[id]/fetchDetails/route.ts
@@ -1,6 +1,6 @@
 // src/app/api/gallery/[id]/description/route.ts
 import { NextResponse } from "next/server";
-import { fetchGalleryDetails } from "@/src/app/lib/db/gallery_tb_utils";
+import { fetchGalleryDetails } from "@/src/app/lib/db/galleryTableUtils";
 
 export async function GET(
   request: Request,
