+++ src/app/lib/db/galleryTableUtils.ts
