+++ src/app/lib/db/splatTableUtils.ts
